Close Issue #273
Issue #273: Memory overrun when text without descenders is rendered with
            underlining

When the fractional top and bottom of an underlining fit within a single row,
the difference between the line bottom and the lower pixel boundary, a
negative number, was misinterpreted as meaning there was a second row to
fill. If the line was already at the bottom row of the surface, then the
next row overwrote invalid memory. This problem was fixed in ft_render_cb.c.

To avoid rounding problems with underlines having an odd thickness value,
changes were also made in ft_render.c.