A possible solution to Issue #131
In surface.c the surf_convert function uses memcpy to copy the source format
to the target format. This includes the palette pointer. If the target format
has a non-NULL palette pointer, the palette is checked to ensure at least one
palette color is not black. It does not matter if the target pixel format
actually uses a palette or not. So unless the target has a palette, make sure
the palette pointer is NULL.