BUG: PixelArray memory leak when using multiple indexes
Py_DECREF was not called on objects returned by PySequence_ITEM(). This was
a problem when 2 subscripts were use for array subcripting—p = array[x, y]—or
array subscript assignment—array[x, y] = p—as neither x nor y could be
garbage collect afterward. It also affected the elements of a sequence
assignmed to an array slice.

Bug reported in email message:
http://www.mail-archive.com/pygame-users@seul.org/msg18966.html

Thanks to mar77i for reporting the bug.